David Fox
Commit failed first time -- trying again

Note: this update retains some bugs in SourceBuffEdSim (formerly in EdSim or
SourceBuff) to simplify regression testing.  Admin\1218bugme.txt is the output
of test all.  It can be compared to test output from earlier versions with
these bugs.  Another update with these bugs fixed will follow immediately.

I have rewritten the source buffer class as an abstract base class, moving
much of the EdSim-specific implementation to a new set class, SourceBuffEdSim.
In the process, I have also made substantial revisions in the SourceBuff and
AppState interface:

1.  Methods instead of attributes

To facilitate connection of future versions of the source buffer to an actual
external editor, there are now methods to access or change the current
position, selection, and contents.  Specifically:

Property		Read method		Change method
------------------------------------------------------------------------------------------
current position		cur_pos()		goto(pos)
selection		get_selection()		set_selection(range, cursor_at)
contents		contents()		insert(...), delete(...), etc.

All access to these properties, except from within a specific concrete
subclass of SourceBuff, should be via these methods, since the underlying
attributes will be implementation-specific.

2. Buffer-specific methods moved to from AppState to SourceBuff

All of the buffer-specific methods from AppState, which took an optional
f_name argument but otherwise defaulted to the current buffer, have been moved
to the source buffer class (without the f_name argument).  These methods are
still accessible from AppState with the optional f_name argument.  Rather than
manually writing a host of AppState methods, each forwarding to the
corresponding source buffer methods, I have added a __getattr__ method to
AppState which catches access to nonexistent attributes of AppState.  If the
attribute is a member of the list AppState.buffer_methods, it is forwarded to
the corresponding method of the specified for current buffer.  Otherwise, an
AttributeError is raised.

Note that new methods added to the source buffer class must have their names
added to AppState.buffer_methods in order to be automatically accessible in
this manner.  Since external access to data from any specific subclass of
AppState should always be through methods declared in the AppState interface,
so as to work with any application, subclasses of AppState should not create
their own buffer_methods attributes or modify AppState.buffer_methods.

3. Each source buffer stores a reference to its containing application in the
app attribute.  App is a required argument to SourceBuff.__init__.  For
SourceBuffEdSim.__init__, I have not enforced app=EdSim, because it is
possible that future simulated editors may find it useful to share the
SourceBuffEdSim implementation.

4.  move_to eliminated

The partially redundant move_to method, which had no range checking, has been
removed.  Use goto(pos), which is range-checked, instead.

5.  Behavior of current selection

For consistency with external editors, when no text is selected, the current
selection is now represented by an empty range coinciding with the current
position.  When the selection is not empty, the current position must always
coincide with one of the end of the selection.  The set_selection method takes
an optional cursor_at argument to specify left (0) or right (1).  Note that
specifying a range with end < start will now result in an empty range, not a
range with the cursor at the left end.

Most Windows editors do not support having the current position at the start
of the current selection, so the AppState method bidirectional_selection (also
accessible from SourceBuff) reports whether or not this is supported.  If it
is not supported, the cursor_at argument is ignored.  Windows and UNIX editors
also have different behavior when texts selected and the left or right arrow
key is pressed.  In UNIX, where the editor keeps track of which end of the
selection corresponds to the current position, the cursor is moved from that
position. In Windows, the cursor is not moved, but is left at the left or
right end of the previous selection range depending on the direction of the
arrow key pressed.  The goto_end_of_selection method provides cross-platform
support for the Windows behavior.

All source buffer methods which change the current position or selection now
maintain the relationship described above.  Note that goto() and any other
methods which change the cursor position (including goto_end_of_selection) or
overwrite the contents of the current selection (including delete and insert),
now make the selection range empty.  To change the cursor position and the
selection, call set_selection, not goto().

Finally, the insert and insert_indent interfaces should now overwrite the
current selection unless the different range is specified.

Since unchecked cursor movement has been eliminated,
EdSim.set_cur_pos_within_bounds has been eliminated.
SourceBuff.make_within_range(pos) remains, since it doesn't deal with the
current position.  However, even SourceBuffEdSim methods should use goto() to
ensure that they never set the current position out of bounds, and to ensure
that the current selection is set consistently.

6.  language_name

The language_name function is now available as an ordinary function in the
SourceBuff module.  Do not create a temporary SourceBuff object merely to
check the association between a filename and a language.

7. new methods:

get_text(start = None, end = None) retrieves a range of the buffer text, with
start and end defaulting to the beginning and end of the buffer, respectively.
contents(), replacing the contents attribute, remains as a synonym for
get_text(), but if only a specific range is desired, get_text(start, end) may
be more efficient than contents()[start:end] with external editors.

len() returns the length of the buffer contents

make_valid_range(range) has been added to ensure that a given range has start
<= end and to truncate it to fall within the buffer's range.  make_valid_range
takes and returns a 2-tuple.

refresh() and refresh_if_need() have been added.  Currently, these do nothing,
but they will eventually provide an application-independent interface to
replace print_buff.

8.  Changes in names and interfaces of a number of buffer-specific methods

The interface of most methods taking a range, currently including
set_selection, insert, indent, insert_indent, and delete, has been modified to
take a single range parameter (2-tuple) rather than separate start and end
attributes.  This facilitates making the current selection the default range.
The exceptions are the new get_text method, where we want start to default to
the beginning of the buffer and end to default to the end of the buffer, with
either or both being optional.

The names of some methods (select, select_region) have also been changed for
consistency with TextBuffer and to eliminate redundant methods.

9.  SourceBuffEdSim reimplemented, fixed

SourceBuffEdSim (formerly simply SourceBuff) has been reimplemented as a
concrete subclass of the now abstract SourceBuff.

10.  All references to the old AppState, EdSim, and SourceBuff methods have
been corrected to use the new interfaces.