Fixed miscellaneous bugs in SourceBuffEdSim (formerly in SourceBuff and EdSim):
- goto_line now goes to an absolute line number, independent of the current
position.

- lines_around_cursor used to incorrectly try to make_within_range to check
whether the starting and ending LINE numbers were within the CHARACTER range
of the buffer.  It now correctly includes +- print_window_size lines.

- print_buff now correctly numbers lines before and after the selection.
(Previously, it often incremented the curr_line_num extra times around the
selection start and end).

- print_buff only prints the start of buffer / end of buffer messages if
at the start and end of the buffer, not if the window is in the middle
of the buffer

- insert and insert_indent now overwrite the current selection unless a
different target range is specified.  (Otherwise, what's the point of a
selection grammar?)