GUI editor demo work.
WaxEdit was renamed to WaxEditDemo
The new WaxEdit is an abstract class which assists in connecting the GUI
editor to the mediator.

WaxEdSim (standalone GUI demo without any voice interface) was renamed to WaxEd (forgot to cvs add - will do this in next commit)

WaxEdSim is the concrete class which is connected to the GUI

gui_sim.py is the GUI equivalent of mediator.py.  It creates a WaxEdSim
and connects it to the MediatorObject.  Unfortunately, it causes a
Windows page fault when recognition starts, so it can't be used by voice
yet.

SourceBuffTB was finished by implementing indent as a no-op.  This needs
to be fixed.

Miscellaneous changes to sim_commands, sr_interface, MediatorObject,
TextBuffer to accomodate the GUI mediator simulator, without messing up
the DOS version still used for regression testing.