More changes/bug-fixes for the basic correction state system.  Most importantly:
1.  Previously, if you dictated something, moved the cursor,
dictated something else in the same buffer, and then corrected the first
utterance, both would be undone and reinterpreted, with the result that
the second utterance would be re-inserted in the wrong place.  A new
StateStackBasic is smart enough to know that it isn't safe to undo and
reinterpret the first utterance.  The problem is only with undoing and
re-interpreting, so you can still say "Scratch Last 2" to
undo both utterances.  Utterances in different windows also don't
interfere with each other.  However, for simplicity, I also made it
illegal to re-interpret any utterance which changes any buffer other
than its initial buffer.

2.  The current implementation of undo using SB_ServiceFullState and
SourceBuffState now has a mix-in class sb_mixins.WithStateService, so
that its methods don't need to be defined explicitly in SourceBuffEdSim,
SourceBuffTB, and SourceBuffMessaging.

3.  I fixed several bugs in the mix-in classes in sb_mixins, and a bug
in OwnerObject which interfered with its system of cleaning up circular
references when a class inherited from multiple OwnerObject classes.

In detail:

messaging:
  - added messarg2inttuple
  - replaced calls to messarg2intlist to convert ranges to lists
    with messarg2inttuple in AppState, SourceBuffMessaging, and tcp_client

SourceBuff, sb_services.SB_ServiceFullState, SourceBuffState:
  - added get_state_pos_selection, compare_state_selections, and
    compare_selection_with_current methods

ResMgr:
  - added changed_buffers method to BufferStates, BufferStatesBasic
  - added selection argument to BufferStates.compare_with_current
  - added safe_reinterp_depth and can_reinterpret to StateStack,
    StateStackBasic
  - created a new StateStackBasic which stores the information necessary to
    tell whether successive utterances in the same buffer were contiguous, and
    prevent reinterpretation of the previous one if they were not

sb_mixins:
  - created a new WithStateService sb_mixin and added it to
    SourceBuff.SourceBuffWithServices
  - removed the explicit SB_ServiceFullState state_srv attributes and methods
    from SourceBuffEdSim, SourceBuffMessaging, SourceBuffTB
  - fixed a bug in the sb_mixins which were failing to initialize the buff
    argument of the SB_Service objects they created by default

Object:
  - added Object.possibly_init_attrs and used it in OwnerObject to fix a bug
    with multiple inheritance from more than one OwnerObject