fixed bug in SourceBuffCached:
insert_cbk was calling _get_text_from_app if it received an
editor-initiated update for an uncached buffer.  This retrieved
the current contents of the buffer *including* any additional changes
which had already been made, but whose update messages had not yet been
processed.  When those subsequent messages were processed, the cache was
corrupted.

The fix is simply to leave the buffer uncached until it is explicitly
synchronized from the mediator end at the next recognition starting.